Flaw #1 Fixed
Usernames weren't checked for duplicates. IF the username is already in the user file it will stay on the Register Page and make the user create a new username because the username is already in the user.txt file. (garettowens123! can't be created because its already in the user.txt commit.)